+++ packages/@aws-cdk/toolkit-lib/lib/api/cloud-assembly/private/context-aware-source.ts
@@ -1,7 +1,7 @@
 import type { MissingContext } from '@aws-cdk/cloud-assembly-schema';
 import * as cxapi from '@aws-cdk/cx-api';
 import { ToolkitServices } from '../../../toolkit/private';
-import { Context, contextproviders, PROJECT_CONTEXT } from '../../aws-cdk';
+import { type Context, contextproviders, PROJECT_CONTEXT } from '../../aws-cdk';
 import { ToolkitError } from '../../errors';
 import { ActionAwareIoHost, debug } from '../../io/private';
 import { ICloudAssemblySource } from '../types';
