+++ packages/aws-cdk/lib/api/deployments/deploy-stack.ts
@@ -36,7 +36,7 @@ import type { SDK, SdkProvider, ICloudFormationClient } from '../aws-auth';
 import type { EnvironmentResources } from '../environment';
 import { CfnEvaluationException } from '../evaluate-cloudformation-template';
 import { HotswapMode, HotswapPropertyOverrides, ICON } from '../hotswap/common';
-import { StackActivityMonitor, type StackActivityProgress } from '../util/cloudformation/stack-activity-monitor';
+import { StackActivityMonitor, type StackActivityProgress } from '../stack-events';
 import { StringWithoutPlaceholders } from '../util/placeholders';
 import { type TemplateBodyParameter, makeBodyParameter } from '../util/template-body-parameter';
