+++ packages/aws-cdk/lib/api/stack-events/display.ts
