+++ packages/aws-cdk/lib/api/stack-events/index.ts
@@ -1,0 +1,3 @@
+export * from './stack-activity-monitor';
+export * from './stack-event-poller';
+export * from './stack-status';
