+++ packages/aws-cdk/lib/cli/cli-config.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { CliHelpers, type CliConfig } from '@aws-cdk/user-input-gen';
 import * as cdk_from_cfn from 'cdk-from-cfn';
-import { StackActivityProgress } from '../api/util/cloudformation/stack-activity-monitor';
+import { StackActivityProgress } from '../api/stack-events';
 import { RequireApproval } from '../diff';
 import { availableInitLanguages } from '../init';
