+++ packages/aws-cdk/test/api/cxapp/cloud-assembly.test.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/order */
 import * as cxschema from '@aws-cdk/cloud-assembly-schema';
-import { DefaultSelection } from '../../lib/api/cxapp/cloud-assembly';
-import { MockCloudExecutable } from '../util';
+import { DefaultSelection } from '../../../lib/api/cxapp/cloud-assembly';
+import { MockCloudExecutable } from '../../util';
 import { cliAssemblyWithForcedVersion } from './assembly-versions';
 
 test('select all top level stacks in the presence of nested assemblies', async () => {
