+++ packages/aws-cdk/test/api/cxapp/exec.test.ts
@@ -6,14 +6,14 @@ import * as cdk from 'aws-cdk-lib';
 import * as semver from 'semver';
 import * as sinon from 'sinon';
 import { ImportMock } from 'ts-mock-imports';
-import { execProgram } from '../../lib/api/cxapp/exec';
-import { Configuration } from '../../lib/cli/user-configuration';
-import { testAssembly } from '../util';
-import { mockSpawn } from '../util/mock-child_process';
-import { MockSdkProvider } from '../util/mock-sdk';
-import { RWLock } from '../../lib/api/util/rwlock';
+import { execProgram } from '../../../lib/api/cxapp/exec';
+import { Configuration } from '../../../lib/cli/user-configuration';
+import { testAssembly } from '../../util';
+import { mockSpawn } from '../../util/mock-child_process';
+import { MockSdkProvider } from '../../util/mock-sdk';
+import { RWLock } from '../../../lib/api/util/rwlock';
 import { rewriteManifestVersion } from './assembly-versions';
-import { CliIoHost } from '../../lib/toolkit/cli-io-host';
+import { CliIoHost } from '../../../lib/toolkit/cli-io-host';
 
 let sdkProvider: MockSdkProvider;
 let config: Configuration;
