+++ packages/aws-cdk/test/api/deployments/cloudformation.test.ts
@@ -1,8 +1,8 @@
 import { SSMPARAM_NO_INVALIDATE } from '@aws-cdk/cx-api';
 import { DescribeStacksCommand, GetTemplateCommand, StackStatus } from '@aws-sdk/client-cloudformation';
-import { MockSdk, mockCloudFormationClient } from './mock-sdk';
-import type { ICloudFormationClient } from '../../lib/api/aws-auth';
-import { CloudFormationStack, TemplateParameters } from '../../lib/api/deployments';
+import { MockSdk, mockCloudFormationClient } from '../../util/mock-sdk';
+import type { ICloudFormationClient } from '../../../lib/api/aws-auth';
+import { CloudFormationStack, TemplateParameters } from '../../../lib/api/deployments';
 
 const PARAM = 'TheParameter';
 const DEFAULT = 'TheDefault';
