+++ packages/aws-cdk/test/api/hotswap/code-build-projects-hotswap-deployments.test.ts
@@ -1,5 +1,5 @@
 import { UpdateProjectCommand } from '@aws-sdk/client-codebuild';
-import * as setup from './hotswap-test-setup';
+import * as setup from '../_helpers/hotswap-test-setup';
 import { HotswapMode } from '../../../lib/api/hotswap/common';
 import { mockCodeBuildClient } from '../../util/mock-sdk';
 import { silentTest } from '../../util/silent';
