+++ packages/aws-cdk/test/api/hotswap/ecs-services-hotswap-deployments.test.ts
@@ -1,5 +1,5 @@
 import { DescribeServicesCommand, RegisterTaskDefinitionCommand, UpdateServiceCommand } from '@aws-sdk/client-ecs';
-import * as setup from './hotswap-test-setup';
+import * as setup from '../_helpers/hotswap-test-setup';
 import { EcsHotswapProperties, HotswapMode, HotswapPropertyOverrides } from '../../../lib/api/hotswap/common';
 import { Configuration } from '../../../lib/cli/user-configuration';
 import { mockECSClient } from '../../util/mock-sdk';
