+++ packages/aws-cdk/test/api/hotswap/lambda-functions-hotswap-deployments.test.ts
@@ -3,7 +3,7 @@ import {
   UpdateFunctionConfigurationCommand,
   waitUntilFunctionUpdatedV2,
 } from '@aws-sdk/client-lambda';
-import * as setup from './hotswap-test-setup';
+import * as setup from '../_helpers/hotswap-test-setup';
 import { HotswapMode } from '../../../lib/api/hotswap/common';
 import { mockLambdaClient } from '../../util/mock-sdk';
 import { silentTest } from '../../util/silent';
