+++ packages/aws-cdk/test/api/hotswap/lambda-versions-aliases-hotswap-deployments.test.ts
@@ -1,5 +1,5 @@
 import { PublishVersionCommand, UpdateAliasCommand } from '@aws-sdk/client-lambda';
-import * as setup from './hotswap-test-setup';
+import * as setup from '../_helpers/hotswap-test-setup';
 import { HotswapMode } from '../../../lib/api/hotswap/common';
 import { mockLambdaClient } from '../../util/mock-sdk';
 import { silentTest } from '../../util/silent';
