+++ packages/aws-cdk/test/api/hotswap/state-machine-hotswap-deployments.test.ts
@@ -1,5 +1,5 @@
 import { UpdateStateMachineCommand } from '@aws-sdk/client-sfn';
-import * as setup from './hotswap-test-setup';
+import * as setup from '../_helpers/hotswap-test-setup';
 import { HotswapMode } from '../../../lib/api/hotswap/common';
 import { mockStepFunctionsClient } from '../../util/mock-sdk';
 import { silentTest } from '../../util/silent';
